+++ jest.config.ts
@@ -97,7 +97,7 @@ const config: Config = {
     '^@store/(.+)$': '<rootDir>/src/app/store/$1',
     '^@home/(.+)$': '<rootDir>/src/app/home/$1',
     '^@about/(.+)$': '<rootDir>/src/app/about/$1',
-    '^@transloco/(.+)$': '<rootDir>/src/app/transloco/$1',
+    '^@internalization/(.+)$': '<rootDir>/src/app/internalization/$1',
 
   },
