+++ src/app/actions/lang.actions.ts
@@ -1,6 +1,6 @@
 import { Source } from '@state-adapt/rxjs';
-import { langStorePrefix } from '@transloco/lang.constants';
-import { Lang } from '@transloco/lang.types';
+import { langStorePrefix } from '@internalization/lang.constants';
+import { Lang } from '@internalization/lang.types';
 
 export const languageChange$ = new Source<Lang>(
   `[${langStorePrefix}] Transloco languageChange$`,
