+++ src/app/app.component.spec.ts
@@ -5,7 +5,7 @@ import { defaultStoreProvider } from '@state-adapt/angular';
 import { AppComponent } from './app.component';
 import { HeaderComponent } from './header/header.component';
 import { ActivatedRoute } from '@angular/router';
-import { getTranslocoTestingModule } from '@transloco/get-transloco-testing-module.function';
+import { getTranslocoTestingModule } from '@internalization/get-transloco-testing-module.function';
 
 describe('AppComponent', () => {
   const fakeActivatedRoute = {
