+++ src/app/internalization/i18n/ro.json
