+++ src/app/internalization/lang-store/lang-store.service.spec.ts
@@ -2,7 +2,7 @@ import { TestBed } from '@angular/core/testing';
 
 import { LangStoreService } from './lang-store.service';
 import { defaultStoreProvider } from '@state-adapt/angular';
-import { getTranslocoTestingModule } from '@transloco/get-transloco-testing-module.function';
+import { getTranslocoTestingModule } from '@internalization/get-transloco-testing-module.function';
 
 describe('LangStoreService', () => {
   let service: LangStoreService;
