+++ src/app/shared/i18n/inline-loader.function.ts
@@ -1,5 +1,5 @@
+import { langList } from '@internalization/lang.constants';
 import { InlineLoader, Translation } from '@jsverse/transloco';
-import { langList } from './lang.constants';
 
 export const inlineLoader = (
   importer: (lang: string, folder: string) => Promise<Translation>,
