+++ tsconfig.json
@@ -26,7 +26,7 @@
       "@store/*": ["src/app/store/*"],
       "@about/*": ["src/app/about/*"],
       "@home/*": ["src/app/home/*"],
-      "@transloco/*": ["src/app/transloco/*"]
+      "@internalization/*": ["src/app/internalization/*"]
 
     }
   },
